refactor(app-core): split plan_apply into finalizers, paths, callbacks, apply, lifecycle modules
Pure move: 4042-line plan_apply.rs becomes a module root retaining the
shared registry/emitter/db_err items, with the rest relocated verbatim
into plan_apply/{finalizers,paths,callbacks,apply,lifecycle,tests}.rs.
The pub mod declaration in lib.rs and every public item path are
unchanged; moved internals gain pub(super) so the root can re-export
them.
